+++ src/main/java/subway/application/SectionService.java
@@ -7,11 +7,11 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.domain.Line;
-import subway.domain.LineRepository;
+import subway.domain.repository.LineRepository;
 import subway.domain.Section;
-import subway.domain.SectionRepository;
+import subway.domain.repository.SectionRepository;
 import subway.domain.Station;
-import subway.domain.StationRepository;
+import subway.domain.repository.StationRepository;
 import subway.dto.section.SectionCreateRequest;
 import subway.dto.section.SectionDeleteRequest;
 import subway.dto.section.SectionResponse;
